Struct ruma_identifiers::RoomAliasId

pub struct RoomAliasId {
    // some fields omitted
}

A Matrix room alias ID.

A RoomAliasId is converted from a string slice, and can be converted back into a string as needed.

assert_eq!(
    RoomAliasId::try_from("#ruma:example.com").unwrap().to_string(),
    "#ruma:example.com"
);

Methods

impl RoomAliasId

fn hostname(&self) -> &Host

Returns a Host for the room alias ID, containing the server name (minus the port) of the originating homeserver.

The host can be either a domain name, an IPv4 address, or an IPv6 address.

fn alias(&self) -> &str

Returns the room's alias.

fn port(&self) -> u16

Returns the port the originating homeserver can be accessed on.