Struct ruma_identifiers::RoomId

pub struct RoomId {
    // some fields omitted
}

A Matrix room ID.

A RoomId is generated randomly or converted from a string slice, and can be converted back into a string as needed.

assert_eq!(
    RoomId::try_from("!n8f893n9:example.com").unwrap().to_string(),
    "!n8f893n9:example.com"
);

Methods

impl RoomId

fn new(server_name: &str) -> Result<Self, Error>

Attempts to generate a RoomId for the given origin server with a localpart consisting of 18 random ASCII characters.

Fails if the given origin server name cannot be parsed as a valid host.

fn hostname(&self) -> &Host

Returns a Host for the room ID, containing the server name (minus the port) of the originating homeserver.

The host can be either a domain name, an IPv4 address, or an IPv6 address.

fn opaque_id(&self) -> &str

Returns the event's opaque ID.

fn port(&self) -> u16

Returns the port the originating homeserver can be accessed on.